[ALPS03975946] Thermal: Fix coverity
[Detail] Unintentional integer overflow
	 (OVERFLOW_BEFORE_WIDEN)

[Solution]cast tpcb_coef to type long long

MTK-Commit-Id: f0603fd7ea2e317e9ca6373cccc779850c7c7618

CR-Id: ALPS03975946
Feature: Thermal Management
(cherry picked from commit d24c7e6dd9809dd9cff6f37483afb352ee6f9e72)

+++ drivers/misc/mediatek/thermal/common/thermal_zones/mtk_ts_dctm.c
@@ -147,8 +147,8 @@ static int tskinInit(int tpcbInit)
 		tpcbInit = tpcbinit;
 
 	mutex_lock(&dctm_mutex);
-	rhs[TSKINNODE] = tamb_coef * tamb; // 10^3
-	rhs[TPCBNODE]  = tpcb_coef * tpcbInit; //10^3
+	rhs[TSKINNODE] = (long long)tamb_coef * tamb; // 10^3
+	rhs[TPCBNODE]  = (long long)tpcb_coef * tpcbInit; //10^3
 
 	for (i = 0; i < NUMNODE; i++) {
 		tInit = 0;
